[CO] Notches for LamHole, bore for Lamination
