+++ pyleecan/Classes/LamSlotMag.py
@@ -69,6 +69,7 @@ from .Slot import Slot
 from .Material import Material
 from .Hole import Hole
 from .Notch import Notch
+from .Bore import Bore
 
 
 class LamSlotMag(LamSlot):
@@ -197,6 +198,7 @@ class LamSlotMag(LamSlot):
         axial_vent=-1,
         notch=-1,
         yoke_notch=-1,
+        bore=None,
         init_dict=None,
         init_str=None,
     ):
@@ -243,6 +245,8 @@ class LamSlotMag(LamSlot):
                 notch = init_dict["notch"]
             if "yoke_notch" in list(init_dict.keys()):
                 yoke_notch = init_dict["yoke_notch"]
+            if "bore" in list(init_dict.keys()):
+                bore = init_dict["bore"]
         # Set the properties (value check and convertion are done in setter)
         self.magnet = magnet
         # Call LamSlot init
@@ -260,6 +264,7 @@ class LamSlotMag(LamSlot):
             axial_vent=axial_vent,
             notch=notch,
             yoke_notch=yoke_notch,
+            bore=bore,
         )
         # The class is frozen (in LamSlot init), for now it's impossible to
         # add new properties
