+++ pyleecan/Classes/LamSquirrelCage.py
@@ -68,6 +68,7 @@ from .Winding import Winding
 from .Slot import Slot
 from .Hole import Hole
 from .Notch import Notch
+from .Bore import Bore
 
 
 class LamSquirrelCage(LamSlotWind):
@@ -192,6 +193,7 @@ class LamSquirrelCage(LamSlotWind):
         axial_vent=-1,
         notch=-1,
         yoke_notch=-1,
+        bore=None,
         init_dict=None,
         init_str=None,
     ):
@@ -246,6 +248,8 @@ class LamSquirrelCage(LamSlotWind):
                 notch = init_dict["notch"]
             if "yoke_notch" in list(init_dict.keys()):
                 yoke_notch = init_dict["yoke_notch"]
+            if "bore" in list(init_dict.keys()):
+                bore = init_dict["bore"]
         # Set the properties (value check and convertion are done in setter)
         self.Hscr = Hscr
         self.Lscr = Lscr
@@ -267,6 +271,7 @@ class LamSquirrelCage(LamSlotWind):
             axial_vent=axial_vent,
             notch=notch,
             yoke_notch=yoke_notch,
+            bore=bore,
         )
         # The class is frozen (in LamSlotWind init), for now it's impossible to
         # add new properties
