+++ pyleecan/Classes/LamSquirrelCageMag.py
@@ -51,6 +51,7 @@ from .Material import Material
 from .Winding import Winding
 from .Slot import Slot
 from .Notch import Notch
+from .Bore import Bore
 
 
 class LamSquirrelCageMag(LamSquirrelCage):
@@ -140,6 +141,7 @@ class LamSquirrelCageMag(LamSquirrelCage):
         axial_vent=-1,
         notch=-1,
         yoke_notch=-1,
+        bore=None,
         init_dict=None,
         init_str=None,
     ):
@@ -196,6 +198,8 @@ class LamSquirrelCageMag(LamSquirrelCage):
                 notch = init_dict["notch"]
             if "yoke_notch" in list(init_dict.keys()):
                 yoke_notch = init_dict["yoke_notch"]
+            if "bore" in list(init_dict.keys()):
+                bore = init_dict["bore"]
         # Set the properties (value check and convertion are done in setter)
         self.hole = hole
         # Call LamSquirrelCage init
@@ -218,6 +222,7 @@ class LamSquirrelCageMag(LamSquirrelCage):
             axial_vent=axial_vent,
             notch=notch,
             yoke_notch=yoke_notch,
+            bore=bore,
         )
         # The class is frozen (in LamSquirrelCage init), for now it's impossible to
         # add new properties
