+++ pyleecan/Classes/Lamination.py
@@ -134,6 +134,7 @@ from ._check import InitUnKnowClassError
 from .Material import Material
 from .Hole import Hole
 from .Notch import Notch
+from .Bore import Bore
 
 
 class Lamination(FrozenClass):
@@ -402,6 +403,7 @@ class Lamination(FrozenClass):
         axial_vent=-1,
         notch=-1,
         yoke_notch=-1,
+        bore=None,
         init_dict=None,
         init_str=None,
     ):
@@ -444,6 +446,8 @@ class Lamination(FrozenClass):
                 notch = init_dict["notch"]
             if "yoke_notch" in list(init_dict.keys()):
                 yoke_notch = init_dict["yoke_notch"]
+            if "bore" in list(init_dict.keys()):
+                bore = init_dict["bore"]
         # Set the properties (value check and convertion are done in setter)
         self.parent = None
         self.L1 = L1
@@ -458,6 +462,7 @@ class Lamination(FrozenClass):
         self.axial_vent = axial_vent
         self.notch = notch
         self.yoke_notch = yoke_notch
+        self.bore = bore
 
         # The class is frozen, for now it's impossible to add new properties
         self._freeze()
@@ -502,6 +507,11 @@ class Lamination(FrozenClass):
                 self.yoke_notch[ii].__str__().replace(linesep, linesep + "\t") + linesep
             )
             Lamination_str += "yoke_notch[" + str(ii) + "] =" + tmp + linesep + linesep
+        if self.bore is not None:
+            tmp = self.bore.__str__().replace(linesep, linesep + "\t").rstrip("\t")
+            Lamination_str += "bore = " + tmp
+        else:
+            Lamination_str += "bore = None" + linesep + linesep
         return Lamination_str
 
     def __eq__(self, other):
@@ -532,6 +542,8 @@ class Lamination(FrozenClass):
         if other.notch != self.notch:
             return False
         if other.yoke_notch != self.yoke_notch:
+            return False
+        if other.bore != self.bore:
             return False
         return True
 
@@ -610,6 +622,12 @@ class Lamination(FrozenClass):
                         other.yoke_notch[ii], name=name + ".yoke_notch[" + str(ii) + "]"
                     )
                 )
+        if (other.bore is None and self.bore is not None) or (
+            other.bore is not None and self.bore is None
+        ):
+            diff_list.append(name + ".bore None mismatch")
+        elif self.bore is not None:
+            diff_list.extend(self.bore.compare(other.bore, name=name + ".bore"))
         return diff_list
 
     def __sizeof__(self):
@@ -634,6 +652,7 @@ class Lamination(FrozenClass):
         if self.yoke_notch is not None:
             for value in self.yoke_notch:
                 S += getsizeof(value)
+        S += getsizeof(self.bore)
         return S
 
     def as_dict(self, **kwargs):
@@ -683,6 +702,10 @@ class Lamination(FrozenClass):
                     Lamination_dict["yoke_notch"].append(obj.as_dict(**kwargs))
                 else:
                     Lamination_dict["yoke_notch"].append(None)
+        if self.bore is None:
+            Lamination_dict["bore"] = None
+        else:
+            Lamination_dict["bore"] = self.bore.as_dict(**kwargs)
         # The class name is added to the dict for deserialisation purpose
         Lamination_dict["__class__"] = "Lamination"
         return Lamination_dict
@@ -703,6 +726,8 @@ class Lamination(FrozenClass):
         self.axial_vent = None
         self.notch = None
         self.yoke_notch = None
+        if self.bore is not None:
+            self.bore._set_None()
 
     def _get_L1(self):
         """getter of L1"""
@@ -982,5 +1007,33 @@ class Lamination(FrozenClass):
         doc=u"""Lamination yoke notches
 
         :Type: [Notch]
+        """,
+    )
+
+    def _get_bore(self):
+        """getter of bore"""
+        return self._bore
+
+    def _set_bore(self, value):
+        """setter of bore"""
+        if isinstance(value, str):  # Load from file
+            value = load_init_dict(value)[1]
+        if isinstance(value, dict) and "__class__" in value:
+            class_obj = import_class("pyleecan.Classes", value.get("__class__"), "bore")
+            value = class_obj(init_dict=value)
+        elif type(value) is int and value == -1:  # Default constructor
+            value = Bore()
+        check_var("bore", value, "Bore")
+        self._bore = value
+
+        if self._bore is not None:
+            self._bore.parent = self
+
+    bore = property(
+        fget=_get_bore,
+        fset=_set_bore,
+        doc=u"""Bore Shape
+
+        :Type: Bore
         """,
     )
