+++ pyleecan/Generator/ClassesRef/Machine/Lamination.csv
@@ -11,7 +11,7 @@ is_stator,-,To fill,0,bool,1,,,,,,get_bore_line,,,
 axial_vent,-,Axial ventilation ducts,0,[Hole],,,,,,,get_Rbo,,,
 notch,-,Lamination bore notches,0,[Notch],,,,,,,get_Ryoke,,,
 yoke_notch,-,Lamination yoke notches,0,[Notch],,,,,,,get_name_phase,,,
-,,,,,,,,,,,plot,,,
+bore,,Bore Shape,,Bore,None,,,,,,plot,,,
 ,,,,,,,,,,,comp_output_geo,,,
 ,,,,,,,,,,,get_polar_eq,,,
 ,,,,,,,,,,,is_outwards,,,
