+++ pyleecan/Methods/Machine/Lamination/get_bore_desc.py
@@ -26,7 +26,9 @@ def get_bore_desc(self, sym=1, line_label=None):
 
     Rbo = self.get_Rbo()
 
-    if self.notch is None or len(self.notch) == 0:
+    if self.bore is not None and self.notch in [None, list()] and sym == 1:
+        return None, self.bore.get_bore_line(label=line_label)
+    elif self.notch is None or len(self.notch) == 0:
         # No notches
         if sym == 1:
             bore_desc = list()
